improved ex01

+++ Common_Core/CPP/Module_07/ex01/iter.hpp
@@ -2,10 +2,11 @@
 #define ITER_HPP
 
 #include <string>
+#include <cstddef>
 
 template <typename T>
-void    iter(T* address, int len, void (*fun)(T const &)) {
-    for (int i = 0; i < len; i++)
+void    iter(T* address, size_t len, void (*fun)(T const &)) {
+    for (size_t i = 0; i < len; i++)
         fun(address[i]);
 }
